refactor(pipeline): move code to utils, so that can be easily reused
